Code cleanup gym obs builders.
