--- tests/ml/ray/environment.yml
@@ -1,8 +1,0 @@
-name: flatland-ray-310
-channels:
-    - defaults
-dependencies:
-    - python=3.10.14
-    - pip=24.2
-    - pip:
-          - -r ${RAY_RUNTIME_ENV_CREATE_WORKING_DIR}/requirements-dev.txt
